+++ contracts/switchboard/native/ArbitrumL2Switchboard.sol
@@ -48,7 +48,7 @@ contract ArbitrumL2Switchboard is NativeSwitchboardBase, INativeReceiver {
     }
 
     function initateNativeConfirmation(bytes32 packetId_) external {
-        uint256 capacitorPacketCount = uint256(uint64(uint256(packetId_)));
+        uint64 capacitorPacketCount = uint64(uint256(packetId_));
         bytes32 root = capacitor__.getRootByCount(capacitorPacketCount);
         if (root == bytes32(0)) revert NoRootFound();
 
@@ -77,7 +77,7 @@ contract ArbitrumL2Switchboard is NativeSwitchboardBase, INativeReceiver {
     function allowPacket(
         bytes32 root_,
         bytes32 packetId_,
-        uint256,
+        uint32,
         uint256
     ) external view override returns (bool) {
         if (tripGlobalFuse) return false;
