+++ contracts/switchboard/native/PolygonL1Switchboard.sol
@@ -31,7 +31,7 @@ contract PolygonL1Switchboard is NativeSwitchboardBase, FxBaseRootTunnel {
      * @param packetId_ - packet id
      */
     function initateNativeConfirmation(bytes32 packetId_) external payable {
-        uint256 capacitorPacketCount = uint256(uint64(uint256(packetId_)));
+        uint64 capacitorPacketCount = uint64(uint256(packetId_));
         bytes32 root = capacitor__.getRootByCount(capacitorPacketCount);
         if (root == bytes32(0)) revert NoRootFound();
 
@@ -56,7 +56,7 @@ contract PolygonL1Switchboard is NativeSwitchboardBase, FxBaseRootTunnel {
     function allowPacket(
         bytes32 root_,
         bytes32 packetId_,
-        uint256,
+        uint32,
         uint256
     ) external view override returns (bool) {
         if (tripGlobalFuse) return false;
