+++ contracts/switchboard/native/PolygonL2Switchboard.sol
@@ -35,7 +35,7 @@ contract PolygonL2Switchboard is NativeSwitchboardBase, FxBaseChildTunnel {
      * @param packetId_ - packet id
      */
     function initateNativeConfirmation(bytes32 packetId_) external payable {
-        uint256 capacitorPacketCount = uint256(uint64(uint256(packetId_)));
+        uint64 capacitorPacketCount = uint64(uint256(packetId_));
         bytes32 root = capacitor__.getRootByCount(capacitorPacketCount);
         if (root == bytes32(0)) revert NoRootFound();
 
@@ -67,7 +67,7 @@ contract PolygonL2Switchboard is NativeSwitchboardBase, FxBaseChildTunnel {
     function allowPacket(
         bytes32 root_,
         bytes32 packetId_,
-        uint256,
+        uint32,
         uint256
     ) external view override returns (bool) {
         if (tripGlobalFuse) return false;
